4.21 ok

+++ part4/blogList/controllers/blogs.js
@@ -60,6 +60,11 @@ blogRouter.post('/', async (req, res) => {
 })
 
 blogRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
   const blog = await Blog.findByIdAndDelete(request.params.id)
   if (!blog) {
     return response.status(404).json({ error: 'Blog not found' })
